fix(utils,containers,net): resolve remaining major review findings
- hdr_histogram.hpp: prevent uint64_t overflow in get_mean() by using
  double accumulator (matches get_std_deviation pattern)
- recorder.hpp: use per-SharedState thread_local map instead of single
  slot, preventing data fragmentation when interleaving recorders
- system_stats.hpp: add compile-time POSIX platform guard
- evicting_queue_bytes.hpp: add total_popped field to Stats, fix
  throughput() to measure actual consumption rate not ID advancement
- transport.hpp: null-check crypto_ before TX-thread derefs, add
  close_requested_ atomic for close_gracefully/stop synchronization

+++ eph-utils/include/eph/utils/hdr_histogram.hpp
@@ -459,18 +459,18 @@ class HdrHistogram {
     [[nodiscard]] double get_mean() const noexcept {
         if (total_count_ == 0) return 0.0;
 
-        uint64_t sum = 0;
+        double sum = 0.0;
         for (int32_t i = 0; i < counts_len_; ++i) {
             uint64_t count = counts_[i];
             if (count > 0) {
                 uint64_t low = value_from_index(i);
                 uint64_t high = next_non_equivalent_value(low);
                 uint64_t mid = low + (high - low) / 2;
-                sum += mid * count;
+                sum += static_cast<double>(mid) * static_cast<double>(count);
             }
         }
 
-        return static_cast<double>(sum) / static_cast<double>(total_count_);
+        return sum / static_cast<double>(total_count_);
     }
 
     [[nodiscard]] double get_std_deviation() const noexcept {
